requirements.txt
aioworkers
aioworkers-aiohttp
pyyaml